+++ src/devices/salus_controls.ts
@@ -1,5 +1,4 @@
 import fz from '../converters/fromZigbee';
-import tz from '../converters/toZigbee';
 import * as exposes from '../lib/exposes';
 import {electricityMeter, onOff} from '../lib/modernExtend';
 import * as ota from '../lib/ota';
@@ -14,35 +13,15 @@ const definitions: DefinitionWithExtend[] = [
         model: 'SPE600',
         vendor: 'Salus Controls',
         description: 'Smart plug (EU socket)',
-        fromZigbee: [fz.on_off, fz.metering],
-        toZigbee: [tz.on_off],
-        configure: async (device, coordinatorEndpoint) => {
-            const endpoint = device.getEndpoint(9);
-            await reporting.bind(endpoint, coordinatorEndpoint, ['genOnOff', 'seMetering']);
-            await reporting.onOff(endpoint);
-            await reporting.instantaneousDemand(endpoint, {min: 5, change: 10});
-            await reporting.currentSummDelivered(endpoint, {min: 5, change: [0, 10]});
-            await endpoint.read('seMetering', ['multiplier', 'divisor']);
-        },
+        extend: [onOff(), electricityMeter({cluster: 'metering'})],
         ota: ota.salus,
-        exposes: [e.switch(), e.power(), e.energy()],
     },
     {
         zigbeeModel: ['SP600'],
         model: 'SP600',
         vendor: 'Salus Controls',
         description: 'Smart plug (UK socket)',
-        fromZigbee: [fz.on_off, fz.SP600_power],
-        exposes: [e.switch(), e.power(), e.energy()],
-        toZigbee: [tz.on_off],
-        configure: async (device, coordinatorEndpoint) => {
-            const endpoint = device.getEndpoint(9);
-            await reporting.bind(endpoint, coordinatorEndpoint, ['genOnOff', 'seMetering']);
-            await reporting.onOff(endpoint);
-            await reporting.instantaneousDemand(endpoint, {min: 5, change: 10});
-            await reporting.currentSummDelivered(endpoint, {min: 5, change: [0, 10]});
-            await endpoint.read('seMetering', ['multiplier', 'divisor']);
-        },
+        extend: [onOff(), electricityMeter({cluster: 'metering', fzMetering: fz.SP600_power})],
         ota: ota.salus,
     },
     {